+++ src/utils/wide_ty.rs
@@ -1,26 +1,31 @@
 use core::ops::{BitAnd, BitOr, BitXor, Not};
 
 use wide::{
-    CmpEq, CmpGt, CmpLt, f32x4, f32x8, f32x16, f64x2, f64x4, f64x8, i8x16, i8x32, i16x8, i16x16,
-    i16x32, i32x4, i32x8, i32x16, i64x2, i64x4, i64x8, u8x16, u8x32, u16x8, u16x16, u16x32, u32x4,
-    u32x8, u32x16, u64x2, u64x4, u64x8,
+    f32x4, f32x8, f32x16, f64x2, f64x4, f64x8, i8x16, i8x32, i16x8, i16x16, i16x32, i32x4, i32x8,
+    i32x16, i64x2, i64x4, i64x8, u8x16, u8x32, u16x8, u16x16, u16x32, u32x4, u32x8, u32x16, u64x2,
+    u64x4, u64x8,
 };
 
 use crate::Scalar;
 
 pub(crate) trait WideTy:
-    Scalar
-    + Not<Output = Self>
-    + BitAnd<Output = Self>
-    + BitOr<Output = Self>
-    + BitXor<Output = Self>
-    + CmpEq<Output = Self>
-    + CmpLt<Output = Self>
-    + CmpGt<Output = Self>
+    Scalar + Not<Output = Self> + BitAnd<Output = Self> + BitOr<Output = Self> + BitXor<Output = Self>
 {
     type Array;
 
     fn new(array: Self::Array) -> Self;
+
+    fn simd_eq(self, other: Self) -> Self;
+
+    fn simd_ne(self, other: Self) -> Self;
+
+    fn simd_lt(self, other: Self) -> Self;
+
+    fn simd_gt(self, other: Self) -> Self;
+
+    fn simd_le(self, other: Self) -> Self;
+
+    fn simd_ge(self, other: Self) -> Self;
 
     fn blend(self, t: Self, f: Self) -> Self;
 
@@ -37,6 +42,36 @@ macro_rules! impl_wide {
             #[inline(always)]
             fn new(array: Self::Array) -> Self {
                 Self::new(array)
+            }
+
+            #[inline(always)]
+            fn simd_eq(self, other: Self) -> Self {
+                self.simd_eq(other)
+            }
+
+            #[inline(always)]
+            fn simd_ne(self, other: Self) -> Self {
+                self.simd_ne(other)
+            }
+
+            #[inline(always)]
+            fn simd_lt(self, other: Self) -> Self {
+                self.simd_lt(other)
+            }
+
+            #[inline(always)]
+            fn simd_gt(self, other: Self) -> Self {
+                self.simd_gt(other)
+            }
+
+            #[inline(always)]
+            fn simd_le(self, other: Self) -> Self {
+                self.simd_le(other)
+            }
+
+            #[inline(always)]
+            fn simd_ge(self, other: Self) -> Self {
+                self.simd_ge(other)
             }
 
             #[inline(always)]